# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server ÖZINŞE.",
        "title": "ÖZINŞE API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "paths": {
...
        "/v1/age-categories": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgeCategories"
                ],
                "summary": "Get AgeCategories List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.AgeCategory"
                            }
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get AgeCategories",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgeCategories"
                ],
                "summary": "Create AgeCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "AgeCategory data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.CreateAgeCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Create AgeCategory",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/age-categories/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgeCategories"
                ],
                "summary": "Get AgeCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.AgeCategory"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get AgeCategory",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgeCategories"
                ],
                "summary": "Update AgeCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "AgeCategory data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.UpdateAgeCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Update AgeCategory",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgeCategories"
                ],
                "summary": "Delete AgeCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ResponseId"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No Permission",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't delete AgeCategory",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.AgeCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "views.CreateAgeCategoryRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
        "views.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "views.ResponseId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "views.UpdateAgeCategoryRequest": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}